week-7/requirements.txt
flask
flask-wtf
wtforms
mysql-connector-python
flask-restful
